Adding Suitcase Fusion 9 support
This is just a duplication of existing recipes changed for Suitcase Fusion 9. They do not seem to have changed anything in regards to how the application is installed. I was able to test the download, pkg, and munki recipes. I do not have access to a JSS to test the jss recipe. Let me know if you need anything else.