Remove redundant SEARCH_URL and SEARCH_PATTERN
The underlying URL changed in the SEARCH_PATTERN which causes the recipe to fail. This was fixed in this commit to the .download recipe here (https://github.com/autopkg/rtrouton-recipes/commit/aea03dad000b90f8da073ce16983f4f1d681ad24), however, since it is defined in this recipe, it overrides the fix.