Reflect change in .app name
The name of the app has changed from `Ultimaker Cura.app` to `Ultimaker-Cura.app`.

This will not resolve #224, however it will would likely complete the run, albeit with a version string of `0.0.0` which would not be desirable.  So....acceptance/merging may(?) should wait until the version string issue is resolved.