Update SketchUp 2025 EN.download.recipe
Download fails with 'No match found on URL' error.  Switching to https://sketchup.trimble.com/en/download/all instead of https://help.sketchup.com/en/downloading-sketchup as the discovery page for the downloads seems to fix that and presumably whatever the issue was that meant you had to download the page first and then process it.