Download URL changed for SketchUp 2026
The download URL for SketchUp 2026 has changed to be a generic redirect from https://sketchup.trimble.com/sketchup/2026/SketchUpPro-dmg to the actual URL for the .dmg.  Updated recipe to reflect this - and removed the URLTextSearcher processor that tried to find the download URL as it always returns the generic redirect URL now (until they change the website again!).